extrayendo la logica del contexto del archivo app
