+++ redux-counter/hook-counter/src/App.jsx
@@ -3,7 +3,6 @@ import './App.css'
 
 import CounterContext from './CounterContext'
 
-import { useReducer } from 'react'
 import { useContext } from 'react'
 
 const Display = () => {
@@ -17,31 +16,16 @@ const Button = ({ type, label }) => {
   return <button onClick={() => dispatch({ type })}>{label}</button>
 }
 
-const counterReducer = (state, action) => {
-  switch (action.type) {
-    case 'INC':
-      return state + 1
-    case 'DEC':
-      return state - 1
-    case 'ZERO':
-      return 0
-    default:
-      return state
-  }
-}
-
 const App = () => {
-  const [counter, counterDispatch] = useReducer(counterReducer, 0)
-
   return (
-    <CounterContext.Provider value={[counter, counterDispatch]}>
+    <>
       <Display />
       <div>
         <Button type="INC" label="+" />
         <Button type="DEC" label="-" />
         <Button type="ZERO" label="0" />
       </div>
-    </CounterContext.Provider>
+    </>
   )
 }
