+++ redux-counter/hook-counter/src/CounterContext.jsx
@@ -1,5 +1,29 @@
-import { createContext } from 'react'
+/* eslint-disable react/prop-types */
+import { createContext, useReducer } from 'react'
+
+const counterReducer = (state, action) => {
+  switch (action.type) {
+    case 'INC':
+      return state + 1
+    case 'DEC':
+      return state - 1
+    case 'ZERO':
+      return 0
+    default:
+      return state
+  }
+}
 
 const CounterContext = createContext()
+
+export const CounterContextProvider = (props) => {
+  const [counter, counterDispatch] = useReducer(counterReducer, 0)
+
+  return (
+    <CounterContext.Provider value={[counter, counterDispatch]}>
+      {props.children}
+    </CounterContext.Provider>
+  )
+}
 
 export default CounterContext
