+++ redux-counter/hook-counter/src/main.jsx
@@ -1,10 +1,11 @@
-import React from 'react'
+import './index.css'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
-import './index.css'
+
+import { CounterContextProvider } from './CounterContext.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
+  <CounterContextProvider>
     <App />
-  </React.StrictMode>,
+  </CounterContextProvider>
 )
